server/requirements.txt
python-dotenv~=1.0.1
flask~=3.0.3
requests~=2.31.0
pydantic~=2.7.1
flask-injector~=0.15.0
gunicorn~=22.0.0
psycopg2-binary~=2.9.9
dynaconf~=3.2.5